refactor : #398 <Auth> 코드 리팩토링
- v2 -> v1으로 통합
- v2 경로 security에서 제거
