+++ com.taken_seat.auth-service/src/main/java/com/taken_seat/auth_service/application/dto/user/v1/UserInfoResponseDtoV1.java
@@ -2,13 +2,8 @@ package com.taken_seat.auth_service.application.dto.user.v1;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.taken_seat.auth_service.application.dto.PageResponseDto;
-import com.taken_seat.auth_service.domain.entity.mileage.Mileage;
-import com.taken_seat.auth_service.domain.entity.user.User;
-import com.taken_seat.auth_service.domain.entity.user.UserCoupon;
 import com.taken_seat.common_service.aop.vo.Role;
-import org.springframework.data.domain.Page;
 
-import java.util.Comparator;
 import java.util.UUID;
 
 @JsonInclude(JsonInclude.Include.NON_NULL) // null 필드 제외
@@ -20,36 +15,4 @@ public record UserInfoResponseDtoV1(
         Role role,
         Integer mileage,
         PageResponseDto<UUID> userCoupons
-){
-
-    public static UserInfoResponseDtoV1 of(User user) {
-        return new UserInfoResponseDtoV1(
-                user.getId(),
-                user.getUsername(),
-                user.getEmail(),
-                user.getPhone(),
-                user.getRole(),
-                null,
-                null
-        );
-    }
-
-    public static UserInfoResponseDtoV1 detailsOf(User user, Page<UserCoupon> userCoupons) {
-        PageResponseDto<UUID> couponsPage = PageResponseDto.of(userCoupons.map(UserCoupon::getCouponId));
-
-        Integer latestMileage = user.getMileages().stream()
-                .max(Comparator.comparing(Mileage::getUpdatedAt))
-                .map(Mileage::getMileage)
-                .orElse(0);
-
-        return new UserInfoResponseDtoV1(
-                user.getId(),
-                user.getUsername(),
-                user.getEmail(),
-                user.getPhone(),
-                user.getRole(),
-                latestMileage,
-                couponsPage
-        );
-    }
-}
+){}
